refactor(utils) put all utils in hooks
