+++ App.tsx
@@ -9,7 +9,7 @@ import GlobalProvider from "~/Contexts/globalContext";
 import theme from "~/Styles/theme.styles";
 import Login from "~/Views/Account/Login";
 import Register from "~/Views/Account/Register";
-import Home from "~/Views/Home";
+import Home from "~/Views/Account/Home";
 import Page from "~/Views/Page";
 import Plus from "~/Views/Plus";
 import Search from "~/Views/Search";
