+++ src/Components/Home/HomeBorne/index.tsx
@@ -33,8 +33,8 @@ import PageContentWrapper from "~/Components/PageContentWrapper";
 import Separator from "~/Components/Ui-kit/Separator";
 import Text from "~/Components/Ui-kit/Text";
 import Theme from "~/Styles/theme.styles";
-import NoEventIllustration from "~/Views/Home/Illustrations/NoEventIllustration";
-import SportIllustration from "~/Views/Home/Illustrations/SportIllustration";
+import NoEventIllustration from "~/Views/Account/Home/Illustrations/NoEventIllustration";
+import SportIllustration from "~/Views/Account/Home/Illustrations/SportIllustration";
 const { color, boxShadow } = Theme;
 
 export default function HomeBorne(): JSX.Element {
