+++ src/Views/Account/Home/Illustrations/NoEventIllustration.tsx
