+++ src/Views/Account/Home/index.tsx
@@ -1,5 +1,5 @@
-import HomeBorne from "../../Components/Home/HomeBorne";
-import HomeMobile from "../../Components/Home/HomeMobile";
+import HomeBorne from "../../../Components/Home/HomeBorne";
+import HomeMobile from "../../../Components/Home/HomeMobile";
 import { useGlobalContext } from "~/Contexts/globalContext";
 
 export default function Home(): JSX.Element {
